Update readme and scripts

+++ controller/clean.sh
@@ -1,4 +1,10 @@
 #!/bin/bash
 
-rm db.sqlite3
-rm -rf */migrations/
+read -r -p "This OBLITERATES all state. Are you sure? [y/N] " response
+if [[ $response =~ ^([yY][eE][sS]|[yY])$ ]]
+then
+	rm db.sqlite3
+	rm -rf */migrations/
+	rm celery-schedule
+fi
+
